# Cache lookups to IEnumerable<IYamlTypeConverter>