--- a/YamlDotNet/Serialization/ObjectGraphVisitors/PreProcessingPhaseObjectGraphVisitorSkeleton.cs
+++ b/YamlDotNet/Serialization/ObjectGraphVisitors/PreProcessingPhaseObjectGraphVisitorSkeleton.cs
@@ -21,6 +21,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using YamlDotNet.Serialization.Utilities;
 
 namespace YamlDotNet.Serialization.ObjectGraphVisitors
@@ -30,16 +31,20 @@ namespace YamlDotNet.Serialization.ObjectGraphVisitors
     /// </summary>
     public abstract class PreProcessingPhaseObjectGraphVisitorSkeleton : IObjectGraphVisitor<Nothing>
     {
-        private readonly TypeConverterCache typeConverters;
+        protected readonly IEnumerable<IYamlTypeConverter> typeConverters; // This is kept for backwards compatibility for subclasses. Use typeConverterCache instead.
+        private readonly TypeConverterCache typeConverterCache;
 
         public PreProcessingPhaseObjectGraphVisitorSkeleton(IEnumerable<IYamlTypeConverter> typeConverters)
         {
-            this.typeConverters = new TypeConverterCache(typeConverters);
+            var tcs = typeConverters?.ToArray() ?? Array.Empty<IYamlTypeConverter>();
+
+            this.typeConverters = tcs;
+            this.typeConverterCache = new TypeConverterCache(tcs);
         }
 
         bool IObjectGraphVisitor<Nothing>.Enter(IPropertyDescriptor? propertyDescriptor, IObjectDescriptor value, Nothing context, ObjectSerializer serializer)
         {
-            if (typeConverters.TryGetConverterForType(value.Type, out _))
+            if (typeConverterCache.TryGetConverterForType(value.Type, out _))
             {
                 return false;
             }