@MattKotsenas 2024-08-16
Alternatively, there aren't many `IYamlTypeConverter`s built-in, and the `Accept()` call is cheap, thus we can get similar benefit by changing the `IEnumerable` to be an array instead and manually unrolling the LINQ calls.

I didn't go that route initially because:

1. It's still a breaking change
2. It's less DRY, as each consumer (and any new consumers in the future) would need to also know to avoid LINQ

If you have a strong preference of either path let me know and I'm happy to make updates here.