@EdwardCooke 2024-09-01
I'm happier with no breaking changes, even in the constructors, but yes, this is public so it can be extended. In fact I saw a consumer of yamldotnet the other day that inherited some of these low level classes and made minor changes. Like the kubernetes client.